% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abcd_stress_test_input}
\alias{abcd_stress_test_input}
\title{generated by running prepare_abcd_data(company_activities, company_emissions, ...) on synthetic data inputs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20922 rows and 11 columns.
}
\source{
see generative script in tests folder
}
\usage{
abcd_stress_test_input
}
\description{
generated by running prepare_abcd_data(company_activities, company_emissions, ...) on synthetic data inputs
}
\keyword{datasets}
