% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{add_column_ald_sector_to_financial_data}
\alias{add_column_ald_sector_to_financial_data}
\title{Title}
\usage{
add_column_ald_sector_to_financial_data(financial_data, companies_data)
}
\arguments{
\item{financial_data}{financial_data}

\item{companies_data}{companies_data}
}
\value{
a dataframe
}
\description{
Title
}
