% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{add_column_company_id_to_eikon_data}
\alias{add_column_company_id_to_eikon_data}
\title{Title}
\usage{
add_column_company_id_to_eikon_data(eikon_data, ids_data)
}
\arguments{
\item{eikon_data}{eikon_data}

\item{ids_data}{ids_data}
}
\value{
a dataframe
}
\description{
Title
}
