% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{add_company_source_info}
\alias{add_company_source_info}
\title{Adds information on where the company information in the processed Eikon data
set comes from.}
\usage{
add_company_source_info(data)
}
\arguments{
\item{data}{A data frame holding processed Eikon data
#'}
}
\value{
A data frame
}
\description{
Adds information on where the company information in the processed Eikon data
set comes from.
}
