% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{add_missing_companies_to_eikon}
\alias{add_missing_companies_to_eikon}
\title{Identify which companies in the masterdata_credit are missing in Eikon}
\usage{
add_missing_companies_to_eikon(data, abcd_data)
}
\arguments{
\item{data}{A data frame holding prepared Eikon data set}

\item{abcd_data}{abcd_data}
}
\description{
Identify which companies in the masterdata_credit are missing in Eikon
}
