% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{aggregate_equity_ownership}
\alias{aggregate_equity_ownership}
\title{aggregate volumes of equity_ownership over duplicated rows}
\usage{
aggregate_equity_ownership(ar_data)
}
\arguments{
\item{ar_data}{ar_data}
}
\description{
aggregate volumes of equity_ownership over duplicated rows
}
