% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{aggregate_financial_indicators}
\alias{aggregate_financial_indicators}
\title{Title}
\usage{
aggregate_financial_indicators(financial_data, grp_cols)
}
\arguments{
\item{financial_data}{financial_data}

\item{grp_cols}{grp_cols}
}
\value{
a dataframe
}
\description{
Title
}
