% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{aggregate_indicator_types}
\alias{aggregate_indicator_types}
\title{Aggregate financial data indicator types using dummies.
This does NOT return dummies as boolean, but as the sum of all indicator
types when grouped by company_id.}
\usage{
aggregate_indicator_types(financial_data, grp_cols)
}
\arguments{
\item{financial_data}{financial_data}

\item{grp_cols}{grp_cols}
}
\value{
a dataframe
}
\description{
Aggregate financial data indicator types using dummies.
This does NOT return dummies as boolean, but as the sum of all indicator
types when grouped by company_id.
}
