% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{aggregate_over_geographies}
\alias{aggregate_over_geographies}
\title{Sum productions and emissions over all other columns.
This is done to aggregate those metrics over the different
countries in a region where a company operates in.}
\usage{
aggregate_over_geographies(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\value{
abcd_data
}
\description{
Sum productions and emissions over all other columns.
This is done to aggregate those metrics over the different
countries in a region where a company operates in.
}
