% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_asset_impact_data.R
\name{aggregate_over_technology_types}
\alias{aggregate_over_technology_types}
\title{Sum production and EF values over all columns except technology type.}
\usage{
aggregate_over_technology_types(ar_data)
}
\arguments{
\item{ar_data}{ar_data}
}
\description{
Sum production and EF values over all columns except technology type.
}
