% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{apply_geographies_renaming_rules}
\alias{apply_geographies_renaming_rules}
\title{Title}
\usage{
apply_geographies_renaming_rules(bench_regions, list_rules)
}
\arguments{
\item{bench_regions}{bench_regions}

\item{list_rules}{list of functions each defining a string renaming rule}
}
\value{
bench_regions
}
\description{
Title
}
