% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_activities}
\alias{company_activities}
\title{GENERATE COMPANY ACTIVITY
Generative operations are repeated for companies assigned to a single sector
and companies assigned to multiple sectors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1586 rows and 11 columns.
}
\source{
generative script in tests folder
}
\usage{
company_activities
}
\description{
Generate random unique company_id identifier
Generate random company name for each company_id
Assign random sectors (single or multiple) to each company. Force companies
being assigned a production in MW to have a production in MWh, and vice-versa.
Assign random iso2c country codes to each company sector
Assign random production values at each row, add random nans, set some random rows to full nans
}
\keyword{datasets}
