% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{compute_financial_averages}
\alias{compute_financial_averages}
\title{Title}
\usage{
compute_financial_averages(
  financial_data,
  minimum_sample_size,
  minimum_ratio_sample,
  allowed_range_npm
)
}
\arguments{
\item{financial_data}{financial_data}

\item{minimum_sample_size}{minimum_sample_size}

\item{minimum_ratio_sample}{minimum_ratio_sample}

\item{allowed_range_npm}{allowed_range_npm}
}
\value{
a dataframe
}
\description{
Title
}
