% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_price_data_long.R
\name{create_automotive_prices}
\alias{create_automotive_prices}
\title{Create automotive prices}
\usage{
create_automotive_prices(Scenarios_AnalysisInput)
}
\arguments{
\item{Scenarios_AnalysisInput}{Scenarios_AnalysisInput}
}
\value{
the automotive price dataframe
}
\description{
Synthetic creation of automotive prices by setting them to 1,
where automotive sectors exist in the scenario data
}
