% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{create_averages_eikon}
\alias{create_averages_eikon}
\title{Create averages of eikon variables for companies with missing values}
\usage{
create_averages_eikon(
  data,
  minimum_sample_size,
  minimum_ratio_sample,
  allowed_range_npm
)
}
\arguments{
\item{data}{A data frame holding processed and possibly filtered Eikon data}

\item{minimum_sample_size}{A numeric vector of length one that determines the
minimum required absolute sample size of the given subgroup. Below this
value we do not calculate an average for this sub group.}

\item{minimum_ratio_sample}{A numeric vector of length one that determines
the minimum required ratio of sub group to overall sample. Below this
value we do not calculate an average for this sub group.}

\item{allowed_range_npm}{A numeric vector indicating the allowed minimum and
maximum values for the net profit margin}
}
\value{
A data frame
}
\description{
This function uses the median for averaging numeric variables
because the results may otherwise be very sensitive to outliers. Averages are
created only based on values we previously obtained directly from eikon.
}
