% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{create_missing_year_rows}
\alias{create_missing_year_rows}
\title{Create new rows replicating the nesting columns for each year of the desired
scope not present.
New rows have NA values for columns outside nesting.}
\usage{
create_missing_year_rows(abcd_data, start_year, time_horizon)
}
\arguments{
\item{abcd_data}{abcd_data}

\item{start_year}{start_year}

\item{time_horizon}{time_horizon}
}
\value{
abcd_data
}
\description{
Create new rows replicating the nesting columns for each year of the desired
scope not present.
New rows have NA values for columns outside nesting.
}
