% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_data.R
\name{create_plan_prod_columns}
\alias{create_plan_prod_columns}
\title{rename columns, and sum ald_production over each company to create
plan_sec_prod column}
\usage{
create_plan_prod_columns(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
rename columns, and sum ald_production over each company to create
plan_sec_prod column
}
