% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_data.R
\name{.sum_or_all_nans}
\alias{.sum_or_all_nans}
\title{Sum values by skipping nans, or return nan if all values are nan
(instead of 0 with sum(... na.rm=T))}
\usage{
.sum_or_all_nans(array)
}
\arguments{
\item{array}{a vector of numbers}
}
\value{
sum of array or nan
}
\description{
Sum values by skipping nans, or return nan if all values are nan
(instead of 0 with sum(... na.rm=T))
}
