% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{drop_always_empty_production}
\alias{drop_always_empty_production}
\title{Drop rows where production is always NA or 0 over the time period}
\usage{
drop_always_empty_production(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
Drop rows where production is always NA or 0 over the time period
}
