% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_data.R
\name{drop_empty_prod_and_ef}
\alias{drop_empty_prod_and_ef}
\title{Drop rows where production or emission are nan}
\usage{
drop_empty_prod_and_ef(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
Drop rows where production or emission are nan
}
