% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{drop_empty_prod_or_ef}
\alias{drop_empty_prod_or_ef}
\title{Drop rows where production OR emission are nan}
\usage{
drop_empty_prod_or_ef(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
Drop rows where production OR emission are nan
}
