% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eikon_data}
\alias{eikon_data}
\title{generated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 593 rows and 8 columns.
}
\source{
see generative script in tests folder
}
\usage{
eikon_data
}
\description{
generated
}
\keyword{datasets}
