% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{fill_empty_years_that_follows}
\alias{fill_empty_years_that_follows}
\title{Fill ald_production and emissions_factor with values of previous years
for a given technology at a company}
\usage{
fill_empty_years_that_follows(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
Fill ald_production and emissions_factor with values of previous years
for a given technology at a company
}
