% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{fill_missing_emission_factor}
\alias{fill_missing_emission_factor}
\title{use avg EFs per ald_business_unit to fill missing values}
\usage{
fill_missing_emission_factor(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
use avg EFs per ald_business_unit to fill missing values
}
