% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_asset_impact_data.R
\name{filter_countries_coverage}
\alias{filter_countries_coverage}
\title{Filter companies by country}
\usage{
filter_countries_coverage(
  ar_data,
  company_informations,
  country_filter = c(),
  filter_hqs = FALSE,
  filter_assets = FALSE
)
}
\arguments{
\item{ar_data}{tibble, Asset Impact data}

\item{company_informations}{tibble, company informations}

\item{country_filter}{character vector, countries to keep in ouput}

\item{filter_hqs}{logical, whether to only keep the (worldwide)
assets whose HeadQuarters are in the country}

\item{filter_assets}{logical, wether to only keep the assets in the country}
}
\value{
tibble, raw Asset Impact data
}
\description{
If filter_hqs and filter_assets are both TRUE,
the function will only rowser where both the asset
and its HeadQuarters are in the country.
}
