% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{filter_sectors_abcd_data}
\alias{filter_sectors_abcd_data}
\title{Filter dataframe to keep only desired sectors}
\usage{
filter_sectors_abcd_data(abcd_data, sector_list)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
Filter dataframe to keep only desired sectors
}
