% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{filter_years_abcd_data}
\alias{filter_years_abcd_data}
\title{filter to keep only desired years
Filtering years and fill/create missing years is done independantly, because
other cleaning steps}
\usage{
filter_years_abcd_data(abcd_data, start_year, time_horizon, additional_year)
}
\arguments{
\item{abcd_data}{abcd_data}

\item{start_year}{start_year}

\item{time_horizon}{time_horizon}

\item{additional_year}{additional_year}
}
\description{
filter to keep only desired years
Filtering years and fill/create missing years is done independantly, because
other cleaning steps
}
