% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{find_missing_companies_credit}
\alias{find_missing_companies_credit}
\title{Identify which companies in the masterdata_credit are missing in Eikon}
\usage{
find_missing_companies_credit(data, eikon_data)
}
\arguments{
\item{data}{A data frame holding the masterdata file in which to find the missings}

\item{eikon_data}{A data frame holding prepared Eikon data set}
}
\description{
Identify which companies in the masterdata_credit are missing in Eikon
}
