% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{find_missing_companies_debt}
\alias{find_missing_companies_debt}
\title{Identify which companies in the masterdata_debt are missing in Eikon}
\usage{
find_missing_companies_debt(data, company_identifier, eikon_data)
}
\arguments{
\item{data}{A data frame holding the masterdata file in which to find the missings}

\item{company_identifier}{A data frame holding combinations of company_id,
company_name and corporate_bond_ticker}

\item{eikon_data}{A data frame holding prepared Eikon data set}
}
\description{
Identify which companies in the masterdata_debt are missing in Eikon
}
