% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{get_all_unique_geographies}
\alias{get_all_unique_geographies}
\title{Title}
\usage{
get_all_unique_geographies(trisk_input_dfs)
}
\arguments{
\item{trisk_input_dfs}{trisk_input_dfs}
}
\value{
all unique geographies accross input dfs
}
\description{
Title
}
