% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_financial_data.R
\name{get_missing_companies_in_financial_data}
\alias{get_missing_companies_in_financial_data}
\title{Title}
\usage{
get_missing_companies_in_financial_data(financial_data, companies_data)
}
\arguments{
\item{financial_data}{financial_data}

\item{companies_data}{companies_data}
}
\value{
a dataframe
}
\description{
Title
}
