% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{group_identical_geographies}
\alias{group_identical_geographies}
\title{GROUP IDENTICAL GEOGRAPHIES}
\usage{
group_identical_geographies(bench_regions, matching_tol)
}
\arguments{
\item{bench_regions}{bench_regions}

\item{matching_tol}{percentage of country matching allowed to group 2 geographies}
}
\description{
GROUP IDENTICAL GEOGRAPHIES
}
