% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{identify_indicator_provenance}
\alias{identify_indicator_provenance}
\title{Aggregate indicator columns to differentiate between raw and averages
Since financial features for 1 company are obtained from multiple countries aggregated,
the "purity" of an indicator is expressed as the ratio of raw values over total (averaged +raw)}
\usage{
identify_indicator_provenance(
  prewrangled_financial_data_stress_test,
  financial_features = c("net_profit_margin", "debt_equity_ratio", "volatility")
)
}
\arguments{
\item{prewrangled_financial_data_stress_test}{prewrangled_financial_data_stress_test}

\item{financial_features}{financial_features}
}
\value{
a dataframe
}
\description{
Aggregate indicator columns to differentiate between raw and averages
Since financial features for 1 company are obtained from multiple countries aggregated,
the "purity" of an indicator is expressed as the ratio of raw values over total (averaged +raw)
}
