% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{keep_available_financial_companies_in_ownership_tree}
\alias{keep_available_financial_companies_in_ownership_tree}
\title{company_id in the ownership tree refer to the company owning the target_company_id}
\usage{
keep_available_financial_companies_in_ownership_tree(
  financial_data,
  ownership_tree
)
}
\arguments{
\item{financial_data}{financial_data}

\item{ownership_tree}{ownership_tree}
}
\value{
a dataframe
}
\description{
company_id in the ownership tree refer to the company owning the target_company_id
}
