% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{load_trisk_inputs}
\alias{load_trisk_inputs}
\title{Title}
\usage{
load_trisk_inputs(
  path_prewrangled_capacity_factors,
  path_price_data_long,
  path_Scenarios_AnalysisInput
)
}
\arguments{
\item{path_prewrangled_capacity_factors}{path_prewrangled_capacity_factors}

\item{path_price_data_long}{path_price_data_long}

\item{path_Scenarios_AnalysisInput}{path_Scenarios_AnalysisInput}
}
\value{
trisk_input_dfs
}
\description{
Title
}
