% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{match_emissions_to_production}
\alias{match_emissions_to_production}
\title{Merge production and emissions data.
Filter rows where the production unit and emission unit match as expected}
\usage{
match_emissions_to_production(company_activities, company_emissions)
}
\arguments{
\item{company_activities}{company_activities dataframe}

\item{company_emissions}{company_emissions dataframe}
}
\description{
Merge production and emissions data.
Filter rows where the production unit and emission unit match as expected
}
