% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{match_location_to_region}
\alias{match_location_to_region}
\title{Title}
\usage{
match_location_to_region(financial_data)
}
\arguments{
\item{financial_data}{financial_data}
}
\value{
a dataframe
}
\description{
Title
}
