% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngfs_carbon_price}
\alias{ngfs_carbon_price}
\title{master ngfs_carbon_price used in the stress test}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1376 rows and 7 columns.
}
\source{
see generative script in tests folder
}
\usage{
ngfs_carbon_price
}
\description{
master ngfs_carbon_price used in the stress test
}
\keyword{datasets}
