% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{pivot_equity_ownership_columns}
\alias{pivot_equity_ownership_columns}
\title{pivot values of Equity Ownership, to be used as yearly production/emissions
TODO when Direct Ownership is not nan, use it when Equity Ownership is nan ?}
\usage{
pivot_equity_ownership_columns(ar_data)
}
\arguments{
\item{ar_data}{ar_data}
}
\description{
pivot values of Equity Ownership, to be used as yearly production/emissions
TODO when Direct Ownership is not nan, use it when Equity Ownership is nan ?
}
