% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{plot_sector_averages}
\alias{plot_sector_averages}
\title{Plot the imputed set of average values by a certain level of granularity}
\usage{
plot_sector_averages(data, x, y, subregion = FALSE)
}
\arguments{
\item{data}{A data frame which contains the created averages for a
combination of sectors and regions}

\item{x}{Variable to display on the x axis}

\item{y}{Variable to display on the y axis}

\item{subregion}{Logical parameter indicates whether to display sub regions
by color or not. Default is FALSE.}
}
\value{
A ggplot2 object
}
\description{
Plot the imputed set of average values by a certain level of granularity
}
