% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{portfolio_data}
\alias{portfolio_data}
\title{generated from company_activities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 9 columns.
}
\source{
generative script in tests folder
}
\usage{
portfolio_data
}
\description{
generated from company_activities
}
\keyword{datasets}
