% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{prepare_abcd_data}
\alias{prepare_abcd_data}
\title{Title}
\usage{
prepare_abcd_data(
  company_activities,
  company_emissions,
  scenarios_geographies,
  start_year,
  time_horizon,
  additional_year,
  sector_list
)
}
\arguments{
\item{company_activities}{company_activities}

\item{company_emissions}{company_emissions}

\item{scenarios_geographies}{scenarios_geographies}

\item{start_year}{start_year}

\item{time_horizon}{time_horizon}

\item{additional_year}{additional_year}

\item{sector_list}{sector_list}
}
\value{
abcd_data
}
\description{
Title
}
