% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_asset_impact_data.R
\name{prepare_asset_impact_data}
\alias{prepare_asset_impact_data}
\title{Prepare Asset Impact data before transformation to abcd}
\usage{
prepare_asset_impact_data(ar_data_path)
}
\arguments{
\item{ar_data_path}{file path to the source Asset Resolution xlsx}
}
\description{
Prepare Asset Impact data before transformation to abcd
}
