% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{prepare_assets_data}
\alias{prepare_assets_data}
\title{Prepare companies emissions and productions dataframes}
\usage{
prepare_assets_data(company_activities, company_emissions)
}
\arguments{
\item{company_activities}{production of companies in different countries.
Production amount is indicated in equity_ownership columns, 1 column per year}

\item{company_emissions}{emissions of companies in different countries.
Emission amount is indicated in equity_ownership columns, 1 column per year}
}
\value{
company_activities : production of companies, 1 row per year
company_emissions : emissions of companies (in tCO2 or tCO2e), 1 row per year
}
\description{
Prepare companies emissions and productions dataframes
}
