% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prewrangled_capacity_factors.R
\name{prepare_capacity_factors_NGFS2021}
\alias{prepare_capacity_factors_NGFS2021}
\title{This function reads in raw ngfs capacity and secondary energy data, as found in the dropbox
under "Raw Data" and creates capacity factors for the power sector.
and prepares these in the format required in the stress test.}
\usage{
prepare_capacity_factors_NGFS2021(data)
}
\arguments{
\item{data}{Tibble that provides raw capacity and secondary energy ngfs data file that is to be
processed}
}
\description{
This function reads in raw ngfs capacity and secondary energy data, as found in the dropbox
under "Raw Data" and creates capacity factors for the power sector.
and prepares these in the format required in the stress test.
}
\seealso{
Other data preparation functions: 
\code{\link{prepare_ngfs_carbon_price}()},
\code{\link{prepare_prewrangled_capacity_factors_WEO2021}()},
\code{\link{prepare_price_data_long_NGFS2021}()},
\code{\link{prepare_price_data_long_WEO2021}()},
\code{\link{prepare_scenario_data}()},
\code{\link{preprepare_ngfs_scenario_data}()}
}
\concept{data preparation functions}
