% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{prepare_eikon_data}
\alias{prepare_eikon_data}
\title{For each company select the final financial value of the best granularity}
\usage{
prepare_eikon_data(
  list_eikon_data,
  security_financial_data,
  consolidated_financial_data,
  prewrangled_ownership_tree,
  abcd_data,
  country_region_bridge,
  n_min_sample,
  min_ratio_sample_subgroup,
  range_profit_margin
)
}
\arguments{
\item{list_eikon_data}{A list of data frames containing the raw input data
from eikon}

\item{security_financial_data}{A data frame containing the
security_financial_data as prepared in the data_preparation repo}

\item{consolidated_financial_data}{A data frame containing the
consolidated_financial_data as prepared in the data_preparation repo}

\item{prewrangled_ownership_tree}{A data frame containing the ownership_tree as provided
by Asset Resolution}

\item{abcd_data}{abcd_data}

\item{country_region_bridge}{A data frame which contains a mapping of
countries to multiple regional levels of interest}

\item{n_min_sample}{A numeric vector of length one, indicating the minimum
required size of the reference subgroups when calculating averages}

\item{min_ratio_sample_subgroup}{A numeric vector of length one, indicating
the minimum required share of the reference subgroups compared to the full
sample size when calculating averages}

\item{range_profit_margin}{A numeric vector indicating the allowed minimum
and maximum values for the net profit margin when calculating the averages}
}
\value{
A data frame
}
\description{
For each company select the final financial value of the best granularity
}
