% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{prepare_eikon_data_input}
\alias{prepare_eikon_data_input}
\title{Prepare raw input data files from Eikon}
\usage{
prepare_eikon_data_input(data)
}
\arguments{
\item{data}{The raw Eikon dataframe}
}
\value{
expected eikon input data
}
\description{
Prepare raw input data files from Eikon
}
