% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_example_data.R
\name{prepare_example_data}
\alias{prepare_example_data}
\title{Prepare example data}
\usage{
prepare_example_data(data)
}
\arguments{
\item{data}{A Tibble holding example data.}
}
\description{
Prepare example data
}
