% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{prepare_financial_data}
\alias{prepare_financial_data}
\title{Title}
\usage{
prepare_financial_data(
  ids_data,
  eikon_data,
  companies_data,
  ownership_tree,
  minimum_sample_size,
  minimum_ratio_sample,
  allowed_range_npm
)
}
\arguments{
\item{ids_data}{ids_data}

\item{eikon_data}{eikon_data}

\item{companies_data}{companies_data}

\item{ownership_tree}{ownership_tree}

\item{minimum_sample_size}{minimum_sample_size}

\item{minimum_ratio_sample}{minimum_ratio_sample}

\item{allowed_range_npm}{allowed_range_npm c(min_npm, max_npm)}
}
\value{
a dataframe
}
\description{
Title
}
