% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scenario_data.R
\name{prepare_geco2023}
\alias{prepare_geco2023}
\title{Function reads GECO2023 raw automotive data and wrangles it into the
required Format.}
\usage{
prepare_geco2023(data)
}
\arguments{
\item{data}{The dataset to be processed. This should be a data frame that
includes the necessary columns for processing.}
}
\value{
A data frame with the processed GECO 2023 data.
}
\description{
Function reads GECO2023 raw automotive data and wrangles it into the
required Format.
}
