% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LCOE_adjusted_price_data.R
\name{prepare_lcoe_adjusted_price_data_oxford2021}
\alias{prepare_lcoe_adjusted_price_data_oxford2021}
\title{This function reads in the price data from Oxford and calculates
the LCOE for the power sector.}
\usage{
prepare_lcoe_adjusted_price_data_oxford2021(
  input_data_lcoe_oxford,
  average_npm_power
)
}
\arguments{
\item{input_data_lcoe_oxford}{Dataet containg Oxford prices}

\item{average_npm_power}{net profit margin in the power sector}
}
\description{
This function reads in the price data from Oxford and calculates
the LCOE for the power sector.
}
