% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LCOE_adjusted_price_data.R
\name{prepare_lcoe_adjusted_price_data_weo}
\alias{prepare_lcoe_adjusted_price_data_weo}
\title{This function reads in the long format prepared price data from IEA scenarios
that contain the LCOE for the power sector. It then derives implied market
prices. This is done by adding the average net profit margin of the 100
largest power companies we can find on top of the costs of the start year.
The absolute margin in that year is then added to all following years as well.
We adjust the price by calculating the cost changes for each technology and
scenario as lcoe_t0 / lcoe_t and applying the resulting cost factor to the
absolute profit margin in each year. This means that if costs increase, the
absolute average profit margin will be smaller over time, whereas falling
costs will increase the absolute profit margin over time.}
\usage{
prepare_lcoe_adjusted_price_data_weo(input_data, average_npm_power, start_year)
}
\arguments{
\item{input_data}{Tibble containing the prepared long format price data with
LCOE for the power sector}

\item{average_npm_power}{net profit margin in the power sector}

\item{start_year}{first year of the scenario}
}
\description{
This function reads in the long format prepared price data from IEA scenarios
that contain the LCOE for the power sector. It then derives implied market
prices. This is done by adding the average net profit margin of the 100
largest power companies we can find on top of the costs of the start year.
The absolute margin in that year is then added to all following years as well.
We adjust the price by calculating the cost changes for each technology and
scenario as lcoe_t0 / lcoe_t and applying the resulting cost factor to the
absolute profit margin in each year. This means that if costs increase, the
absolute average profit margin will be smaller over time, whereas falling
costs will increase the absolute profit margin over time.
}
