% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ngfs_carbon_price.R
\name{prepare_ngfs_carbon_price}
\alias{prepare_ngfs_carbon_price}
\title{This function reads carbon price data sourced from the NGFS scenario selector
tool and prepares it for use in the r2dii.climate.stress.test package. Gaps
between years are interpolated.}
\usage{
prepare_ngfs_carbon_price(data, start_year)
}
\arguments{
\item{data}{Tibble that contains the raw ngfs scenario carbon price data file
that is to be processed}

\item{start_year}{Numeric vector of length 1. Indicates the desired start
year of the prepared ngfs price data set}
}
\description{
This function reads carbon price data sourced from the NGFS scenario selector
tool and prepares it for use in the r2dii.climate.stress.test package. Gaps
between years are interpolated.
}
\seealso{
Other data preparation functions: 
\code{\link{prepare_capacity_factors_NGFS2021}()},
\code{\link{prepare_prewrangled_capacity_factors_WEO2021}()},
\code{\link{prepare_price_data_long_NGFS2021}()},
\code{\link{prepare_price_data_long_WEO2021}()},
\code{\link{prepare_scenario_data}()},
\code{\link{preprepare_ngfs_scenario_data}()}
}
\concept{data preparation functions}
