% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prewrangled_capacity_factors.R
\name{prepare_prewrangled_capacity_factors_WEO2021}
\alias{prepare_prewrangled_capacity_factors_WEO2021}
\title{This function reads scenario data in the form of data as found in the dropbox
under "Raw Data", extracts the relevant information for power capacity factors
and prepares these in the format required in the stress test. Currently
expects to get data for up to year 2040.}
\usage{
prepare_prewrangled_capacity_factors_WEO2021(data, start_year)
}
\arguments{
\item{data}{Tibble that provides raw scenario data file that is to be}

\item{start_year}{Beginning of the analysis.
processed}
}
\description{
This function reads scenario data in the form of data as found in the dropbox
under "Raw Data", extracts the relevant information for power capacity factors
and prepares these in the format required in the stress test. Currently
expects to get data for up to year 2040.
}
\seealso{
Other data preparation functions: 
\code{\link{prepare_capacity_factors_NGFS2022}()},
\code{\link{prepare_ngfs_carbon_price}()},
\code{\link{prepare_price_data_long_NGFS2022}()},
\code{\link{prepare_price_data_long_WEO2021}()},
\code{\link{prepare_scenario_data}()},
\code{\link{preprepare_ngfs_scenario_data}()}
}
\concept{data preparation functions}
