% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prewrangled_capacity_factors.R
\name{prepare_prewrangled_capacity_factors_WEO2023}
\alias{prepare_prewrangled_capacity_factors_WEO2023}
\title{WEO2023
This function reads prewrangled weo2023 capacity and net generation data for APS, STEPS and NZ scenarios
under the Global geography}
\usage{
prepare_prewrangled_capacity_factors_WEO2023(data, start_year)
}
\arguments{
\item{data}{A dataframe containing the preprocessed data that needs to be wrangled.
Expected to have columns for sector, indicator, value, source,
scenario, scenario_geography, technology, year, and units.}

\item{start_year}{The starting year from which capacity factors should be calculated.}
}
\description{
WEO2023
This function reads prewrangled weo2023 capacity and net generation data for APS, STEPS and NZ scenarios
under the Global geography
}
