% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{prepare_prewrangled_financial_data_stress_test}
\alias{prepare_prewrangled_financial_data_stress_test}
\title{Prewrangle the prepared eikon data into the format used by the stress test}
\usage{
prepare_prewrangled_financial_data_stress_test(eikon_data)
}
\arguments{
\item{eikon_data}{A data frame which contains the prepared eikon data as
provided after running prepare_eikon_data()}
}
\value{
A list of three data.frames
}
\description{
Prewrangle the prepared eikon data into the format used by the stress test
}
