% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_price_data_long.R
\name{prepare_price_data_long_WEO2021}
\alias{prepare_price_data_long_WEO2021}
\title{This function reads in multiple price data files from WEO2021 and wrangles
it to fit the format to be used in the transition risk stress test work flow.}
\usage{
prepare_price_data_long_WEO2021(input_data_fossil_fuel, input_data_power)
}
\arguments{
\item{input_data_fossil_fuel}{Tibble containing the raw price data for fossil
fuels}

\item{input_data_power}{Tibble containing the raw price/LCOE data for power}
}
\description{
This function reads in multiple price data files from WEO2021 and wrangles
it to fit the format to be used in the transition risk stress test work flow.
}
\seealso{
Other data preparation functions: 
\code{\link{prepare_capacity_factors_NGFS2022}()},
\code{\link{prepare_ngfs_carbon_price}()},
\code{\link{prepare_prewrangled_capacity_factors_WEO2021}()},
\code{\link{prepare_price_data_long_NGFS2022}()},
\code{\link{prepare_scenario_data}()},
\code{\link{preprepare_ngfs_scenario_data}()}
}
\concept{data preparation functions}
