% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scenario_data.R
\name{prepare_scenario_data}
\alias{prepare_scenario_data}
\title{This function reads scenario data in the form of data as found in the dropbox
under "Processed Data" and wrangles it to fit the required format for the
stress test.}
\usage{
prepare_scenario_data(data, start_year)
}
\arguments{
\item{data}{Tibble that contains the scenario data file that is to be
processed}

\item{start_year}{Numeric holding start year.}
}
\description{
This function reads scenario data in the form of data as found in the dropbox
under "Processed Data" and wrangles it to fit the required format for the
stress test.
}
\seealso{
Other data preparation functions: 
\code{\link{prepare_capacity_factors_NGFS2021}()},
\code{\link{prepare_ngfs_carbon_price}()},
\code{\link{prepare_prewrangled_capacity_factors_WEO2021}()},
\code{\link{prepare_price_data_long_WEO2021}()},
\code{\link{preprepare_ngfs_scenario_data}()}
}
\concept{data preparation functions}
