% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scenario_data.R
\name{preprepare_ngfs_scenario_data}
\alias{preprepare_ngfs_scenario_data}
\title{This function reads  NGFS raw scenario data as found in the dropbox
under "Processed Data" and wrangles it to fit the required format for the
usual scenario analysis input routine.}
\usage{
preprepare_ngfs_scenario_data(data)
}
\arguments{
\item{data}{Tibble that contains the scenario data file that is to be
processed}
}
\description{
This function reads  NGFS raw scenario data as found in the dropbox
under "Processed Data" and wrangles it to fit the required format for the
usual scenario analysis input routine.
}
\seealso{
Other data preparation functions: 
\code{\link{prepare_capacity_factors_NGFS2021}()},
\code{\link{prepare_ngfs_carbon_price}()},
\code{\link{prepare_prewrangled_capacity_factors_WEO2021}()},
\code{\link{prepare_price_data_long_NGFS2021}()},
\code{\link{prepare_price_data_long_WEO2021}()},
\code{\link{prepare_scenario_data}()}
}
\concept{data preparation functions}
