% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{prewrangle_eikon_data}
\alias{prewrangle_eikon_data}
\title{Prewrangle the input data files from Eikon}
\usage{
prewrangle_eikon_data(data, security_financial_data_input)
}
\arguments{
\item{data}{A data frame holding prepared Eikon input data sets}

\item{security_financial_data_input}{A data frame holding
security_financial_data}
}
\description{
Prewrangle the input data files from Eikon
}
