% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{prewrangle_masterdata_credit}
\alias{prewrangle_masterdata_credit}
\title{Prewrangle the raw master data credit input file from AR}
\usage{
prewrangle_masterdata_credit(masterdata_credit, consolidated_financial_data)
}
\arguments{
\item{masterdata_credit}{A data frame holding the raw master data credit data}

\item{consolidated_financial_data}{A data frame holding
consolidated_financial_data}
}
\description{
Prewrangle the raw master data credit input file from AR
}
