% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{prewrangle_ownership_tree}
\alias{prewrangle_ownership_tree}
\title{Prewrangle the ownership tree data from AR}
\usage{
prewrangle_ownership_tree(data)
}
\arguments{
\item{data}{A data frame holding the raw ownership tree data set}
}
\description{
Prewrangle the ownership tree data from AR
}
