% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prewrangled_capacity_factors}
\alias{prewrangled_capacity_factors}
\title{master prewrangled_capacity_factors used in the stress test}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 119506 rows and 5 columns.
}
\source{
see generative script in tests folder
}
\usage{
prewrangled_capacity_factors
}
\description{
master prewrangled_capacity_factors used in the stress test
}
\keyword{datasets}
