% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prewrangled_financial_data_stress_test}
\alias{prewrangled_financial_data_stress_test}
\title{generated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 6 columns.
}
\source{
generative script in tests folder
}
\usage{
prewrangled_financial_data_stress_test
}
\description{
generated
}
\keyword{datasets}
