% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_data_long}
\alias{price_data_long}
\title{master price_data_long used in the stress test}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18493 rows and 8 columns.
}
\source{
see generative script in tests folder
}
\usage{
price_data_long
}
\description{
master price_data_long used in the stress test
}
\keyword{datasets}
