% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production_types}
\alias{production_types}
\title{Available production types in the stress test}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 4 columns.
}
\source{
internal
}
\usage{
production_types
}
\description{
Available production types in the stress test
}
\keyword{datasets}
