% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_data.R
\name{read_asset_resolution}
\alias{read_asset_resolution}
\title{read Asset Resolution data}
\usage{
read_asset_resolution(path_ar_data_raw, sheet_name)
}
\arguments{
\item{path_ar_data_raw}{path to AR excel input}

\item{sheet_name}{name of excel sheet}
}
\description{
read Asset Resolution data
}
