% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{regroup_and_rename_geographies}
\alias{regroup_and_rename_geographies}
\title{Title}
\usage{
regroup_and_rename_geographies(
  bench_regions,
  path_prewrangled_capacity_factors,
  path_price_data_long,
  path_Scenarios_AnalysisInput,
  matching_tol = 1
)
}
\arguments{
\item{bench_regions}{bench_regions}

\item{path_prewrangled_capacity_factors}{path_prewrangled_capacity_factors}

\item{path_price_data_long}{path_price_data_long}

\item{path_Scenarios_AnalysisInput}{path_Scenarios_AnalysisInput}

\item{matching_tol}{percentage of country matching allowed to group 2 geographies}
}
\value{
list containing bench_regions and trisk_input_dfs
}
\description{
Title
}
