% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{remove_implausible_values_in_financial_indicators}
\alias{remove_implausible_values_in_financial_indicators}
\title{Lower or equal for min value (0) Stricly higher for the max value (npm==1 is allowed)}
\usage{
remove_implausible_values_in_financial_indicators(
  financial_data,
  allowed_range_npm
)
}
\arguments{
\item{financial_data}{financial_data}

\item{allowed_range_npm}{allowed_range_npm}
}
\value{
a dataframe
}
\description{
Lower or equal for min value (0) Stricly higher for the max value (npm==1 is allowed)
}
