% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_data.R
\name{remove_prop_emissions}
\alias{remove_prop_emissions}
\title{Filtering emissions to remove emissions in proportions
in order to aggregate the raw values and re-compute the proportions}
\usage{
remove_prop_emissions(company_emissions)
}
\arguments{
\item{company_emissions}{company_emissions}
}
\description{
Filtering emissions to remove emissions in proportions
in order to aggregate the raw values and re-compute the proportions
}
