% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{remove_unknown_owner_companies}
\alias{remove_unknown_owner_companies}
\title{Filter out companies with unknown owner}
\usage{
remove_unknown_owner_companies(ar_data)
}
\arguments{
\item{ar_data}{ar_data}
}
\description{
Filter out companies with unknown owner
}
