% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{rename_ald_sector}
\alias{rename_ald_sector}
\title{rename ald_sector column according to some rules}
\usage{
rename_ald_sector(ar_data)
}
\arguments{
\item{ar_data}{ar_data}
}
\description{
rename ald_sector column according to some rules
}
