% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{rename_bench_region_geographies}
\alias{rename_bench_region_geographies}
\title{apply a function to rename a vector of characters, and apply the mapping on the original column}
\usage{
rename_bench_region_geographies(bench_regions, renaming_fun)
}
\arguments{
\item{bench_regions}{bench_regions}

\item{renaming_fun}{a renaming function that takes a single string as parameter}
}
\description{
apply a function to rename a vector of characters, and apply the mapping on the original column
}
