% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rename_column_values}
\alias{rename_column_values}
\title{Apply a rename mapping on a tibble column}
\usage{
rename_column_values(data, colname, key_value_mapping)
}
\arguments{
\item{data}{A tibble}

\item{colname}{name of the column with values to be renamed}

\item{key_value_mapping}{A named vector describing the mapping, keys as old name, values as new name}
}
\value{
input dataframe \code{data} with \code{colname} column values renamed according to \code{key_value_mapping}
}
\description{
Apply a rename mapping on a tibble column
}
