% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{rename_stress_test_inputs}
\alias{rename_stress_test_inputs}
\title{Title}
\usage{
rename_stress_test_inputs(bench_regions, trisk_input_dfs)
}
\arguments{
\item{bench_regions}{bench_regions}

\item{trisk_input_dfs}{trisk_input_dfs}
}
\value{
trisk_input_dfs
}
\description{
Title
}
