% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_data.R
\name{rename_technology}
\alias{rename_technology}
\title{rename technology column according to some rules}
\usage{
rename_technology(ar_data)
}
\arguments{
\item{ar_data}{ar_data}
}
\description{
rename technology column according to some rules
}
