% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{report_diff_rows}
\alias{report_diff_rows}
\title{Identify how many rows are added or removed due to a specific operation}
\usage{
report_diff_rows(data, initial_n_rows, cause = "")
}
\arguments{
\item{data}{A data frame for which to indicate differences in the number of
rows pre and post an operation}

\item{initial_n_rows}{A numeric vector of length one that indicates the
number of rows before the application of the operation at hand. Usually,
this can get nrow(data) as an input.}

\item{cause}{A string that clarifies the reason for a difference in rows for
the application at hand.}
}
\description{
Identify how many rows are added or removed due to a specific operation
}
