% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{report_duplicates}
\alias{report_duplicates}
\title{Report duplicate rows}
\usage{
report_duplicates(data, cols, throw_error = TRUE)
}
\arguments{
\item{data}{Tibble holding a data set.}

\item{cols}{Cols to check for duplicate combinations on.}

\item{throw_error}{Boolean, indicating if error shall be thrown if missings
are detected.}
}
\description{
Reports duplicates in \code{data} on columns \code{cols}. Duplicates are reported via a
warning.
}
