% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{report_missing_col_combinations}
\alias{report_missing_col_combinations}
\title{Identify and report missing value combinations}
\usage{
report_missing_col_combinations(data, cols, throw_error = FALSE)
}
\arguments{
\item{data}{Tibble holding a data set.}

\item{cols}{Vector holding name of columns}

\item{throw_error}{Boolean, indicating if error shall be thrown if missings
are detected.}
}
\description{
Checks if all level combinations of \code{cols} are in\code{data} and
throws a warning on missing combinations.
NOTE:
\enumerate{
\item a combination of all levels is not necessarily required/useful, make sure
to use function only in adequate context.
\item combinations of too many columns/values may exceed memory size.
.
}
}
