% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{report_missings}
\alias{report_missings}
\title{Report missing}
\usage{
report_missings(data, name_data, throw_error = TRUE)
}
\arguments{
\item{data}{Tibble holding a data set.}

\item{name_data}{Name of the data file.}

\item{throw_error}{Boolean, indicating if error shall be thrown if missings
are detected.}
}
\value{
input \code{data}.
}
\description{
Function reports number of missing values per variable.
}
