% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{rm_companies_without_abcd}
\alias{rm_companies_without_abcd}
\title{Removes companies from processed eikon data for which no asset based company
data can be found for any asset type.}
\usage{
rm_companies_without_abcd(data, md_ownership, md_debt, md_credit)
}
\arguments{
\item{data}{A data frame holding processed Eikon data}

\item{md_ownership}{A data frame containing the masterdata_ownership data set}

\item{md_debt}{A data frame containing the masterdata_debt data set}

\item{md_credit}{A data frame containing the masterdata_credit data set}
}
\value{
A data frame
}
\description{
Removes companies from processed eikon data for which no asset based company
data can be found for any asset type.
}
