% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Scenarios_AnalysisInput}
\alias{Scenarios_AnalysisInput}
\title{master Scenarios_AnalysisInput used in the stress test}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 161980 rows and 9 columns.
}
\source{
see generative script in tests folder
}
\usage{
Scenarios_AnalysisInput
}
\description{
master Scenarios_AnalysisInput used in the stress test
}
\keyword{datasets}
