% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenarios_geographies}
\alias{scenarios_geographies}
\title{available geographies in the stress test}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2509 rows and 5 columns.
}
\source{
internal
}
\usage{
scenarios_geographies
}
\description{
available geographies in the stress test
}
\examples{
scenarios_geographies
}
\keyword{datasets}
