% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{select_final_financial_value}
\alias{select_final_financial_value}
\title{For each company select the final financial value of the best granularity}
\usage{
select_final_financial_value(data)
}
\arguments{
\item{data}{A data frame which contains the financial data including the
calculated averages for a combination of sectors and regions}
}
\value{
A data frame
}
\description{
For each company select the final financial value of the best granularity
}
