% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{select_final_financial_value_using_averages}
\alias{select_final_financial_value_using_averages}
\title{For each company select the final financial value of the best granularity}
\usage{
select_final_financial_value_using_averages(financial_data)
}
\arguments{
\item{financial_data}{A data frame which contains the financial data including the
calculated averages for a combination of sectors and regions}
}
\value{
A data frame
}
\description{
For each company select the final financial value of the best granularity
}
