% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geographies.R
\name{standardize_geoographies_name}
\alias{standardize_geoographies_name}
\title{Title}
\usage{
standardize_geoographies_name(bench_regions)
}
\arguments{
\item{bench_regions}{bench_regions}
}
\value{
bench_regions
}
\description{
Title
}
