% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_company_emissions}
\alias{synthetic_company_emissions}
\title{GENERATE COMPANY EMISSIONS
generated from company_activities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1737 rows and 11 columns.
}
\source{
see generative script in tests folder
}
\usage{
synthetic_company_emissions
}
\description{
GENERATE COMPANY EMISSIONS
generated from company_activities
}
\keyword{datasets}
