% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{add_columns_financial_averages}
\alias{add_columns_financial_averages}
\title{Title}
\usage{
add_columns_financial_averages(
  financial_data,
  ald_sector_region_averages,
  ald_sector_averages,
  global_averages
)
}
\arguments{
\item{financial_data}{financial_data}

\item{ald_sector_region_averages}{ald_sector_region_averages}

\item{ald_sector_averages}{ald_sector_averages}

\item{global_averages}{global_averages}
}
\value{
a dataframe
}
\description{
Title
}
