% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eikon_data.R
\name{add_missing_companies_to_eikon}
\alias{add_missing_companies_to_eikon}
\title{Identify which companies in the masterdata_credit are missing in Eikon}
\usage{
add_missing_companies_to_eikon(
  data,
  missing_companies_ownership,
  missing_companies_debt,
  missing_companies_credit
)
}
\arguments{
\item{data}{A data frame holding prepared Eikon data set}

\item{missing_companies_ownership}{A data frame with missing companies from
the masterdata_ownership data set}

\item{missing_companies_debt}{A data frame with missing companies from
the masterdata_debt data set}

\item{missing_companies_credit}{A data frame with missing companies from
the masterdata_credit data set}
}
\description{
Identify which companies in the masterdata_credit are missing in Eikon
}
