% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{aggregate_over_technology_types}
\alias{aggregate_over_technology_types}
\title{Sum production and EF values over all columns except technology type.}
\usage{
aggregate_over_technology_types(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
Sum production and EF values over all columns except technology type.
}
