% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{expand_by_scenario_geography}
\alias{expand_by_scenario_geography}
\title{Duplicate rows according to the available geographies}
\usage{
expand_by_scenario_geography(
  abcd_data,
  scenarios_geographies,
  .default = "Global",
  .iso2c = "ald_location"
)
}
\arguments{
\item{abcd_data}{abcd_data}

\item{scenarios_geographies}{scenarios_geographies}

\item{.default}{.default}

\item{.iso2c}{.iso2c}
}
\description{
Duplicate rows according to the available geographies
}
