% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{fill_partially_missing_values}
\alias{fill_partially_missing_values}
\title{Fill ald_production and emissions_factor for a given technology at a company
with values of previous years, or with interpolation for values in the middle.}
\usage{
fill_partially_missing_values(abcd_data)
}
\arguments{
\item{abcd_data}{abcd_data}
}
\description{
Fill ald_production and emissions_factor for a given technology at a company
with values of previous years, or with interpolation for values in the middle.
}
