% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{filter_years_abcd_data}
\alias{filter_years_abcd_data}
\title{#' Sum production values for technologies that have the same name, after
#' for example a renaming has been applied
#' @param abcd_data
#'
#' @export
aggregate_production_after_renaming <- function(abcd_data) {
abcd_data <- abcd_data \%>\%
dplyr::group_by(
.data$id,
.data$company_name,
.data$ald_sector,
.data$technology,
.data$technology_type,
.data$region,
.data$ald_location,
.data$ald_production_unit,
.data$year
) \%>\%
dplyr::summarise(ald_production = sum(.data$ald_production, na.rm = T))
return(abcd_data)
}
filter to keep only desired years}
\usage{
filter_years_abcd_data(abcd_data, start_year, time_horizon, additional_year)
}
\arguments{
\item{abcd_data}{abcd_data}

\item{start_year}{start_year}

\item{time_horizon}{time_horizon}

\item{additional_year}{additional_year}
}
\description{
#' Sum production values for technologies that have the same name, after
#' for example a renaming has been applied
#' @param abcd_data
#'
#' @export
aggregate_production_after_renaming <- function(abcd_data) {
abcd_data <- abcd_data \%>\%
dplyr::group_by(
.data$id,
.data$company_name,
.data$ald_sector,
.data$technology,
.data$technology_type,
.data$region,
.data$ald_location,
.data$ald_production_unit,
.data$year
) \%>\%
dplyr::summarise(ald_production = sum(.data$ald_production, na.rm = T))
return(abcd_data)
}
filter to keep only desired years
}
