% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{get_missing_companies_in_financial_data}
\alias{get_missing_companies_in_financial_data}
\title{Anti join the financial data with companies data to create new rows
in the financial data for the missing companies.}
\usage{
get_missing_companies_in_financial_data(financial_data, companies_data)
}
\arguments{
\item{financial_data}{financial_data}

\item{companies_data}{companies_data}
}
\value{
a dataframe
}
\description{
Anti join the financial data with companies data to create new rows
in the financial data for the missing companies.
}
