% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{match_closest_financial_data_to_missing_companies}
\alias{match_closest_financial_data_to_missing_companies}
\title{Use the ownership tree to assign raw eikon financial data to a subsidiary company when it
doesn't own any financial information in the raw data}
\usage{
match_closest_financial_data_to_missing_companies(
  missing_companies_in_financial_data,
  financial_data,
  ownership_tree
)
}
\arguments{
\item{missing_companies_in_financial_data}{missing_companies_in_financial_data}

\item{financial_data}{financial_data}

\item{ownership_tree}{ownership_tree}
}
\value{
a dataframe
}
\description{
Use the ownership tree to assign raw eikon financial data to a subsidiary company when it
doesn't own any financial information in the raw data
}
