% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{prepare_abcd_data}
\alias{prepare_abcd_data}
\title{Creates the abcd stress test input data, using companies emissions and productions.
Aggregate absolute values of production and emissions over ald_business_unit types and regions.
Computes the emissions factor ratio out of absolute emissions and productions.}
\usage{
prepare_abcd_data(
  company_activities,
  company_emissions,
  scenarios_geographies,
  start_year,
  time_horizon,
  additional_year,
  km_per_vehicle,
  sector_list
)
}
\arguments{
\item{company_activities}{production of companies in different countries.
Production amount is indicated in equity_ownership columns, 1 column per year}

\item{company_emissions}{emissions of companies in different countries.
Emission amount is indicated in equity_ownership columns, 1 column per year}

\item{scenarios_geographies}{mapping between country code and climate scenario geographies}

\item{start_year}{start_year}

\item{time_horizon}{time_horizon}

\item{additional_year}{list of years to add to the year range}

\item{km_per_vehicle}{the number of km associated to the production of 1 vehicle,
used to convert the emissions from TCO2/# vehicles to TCO2/km}

\item{sector_list}{sector_list}
}
\value{
companies production matched to the appropriate emission, 1 row per year and scenario geography
}
\description{
Creates the abcd stress test input data, using companies emissions and productions.
Aggregate absolute values of production and emissions over ald_business_unit types and regions.
Computes the emissions factor ratio out of absolute emissions and productions.
}
