% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_abcd_stress_test_input.R
\name{prepare_assets_data}
\alias{prepare_assets_data}
\title{Title}
\usage{
prepare_assets_data(company_activities, company_emissions)
}
\arguments{
\item{company_activities}{company_activities}

\item{company_emissions}{company_emissions}
}
\value{
company_activities, company_emissions
}
\description{
Title
}
