% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_prewrangled_financial_data_stress_test.R
\name{prepare_financial_data}
\alias{prepare_financial_data}
\title{Main financial data preparation script}
\usage{
prepare_financial_data(
  financial_data,
  companies_data,
  ownership_tree,
  minimum_sample_size,
  minimum_ratio_sample,
  allowed_range_npm
)
}
\arguments{
\item{financial_data}{financial_data}

\item{companies_data}{companies_data}

\item{ownership_tree}{ownership_tree}

\item{minimum_sample_size}{minimum_sample_size}

\item{minimum_ratio_sample}{minimum_ratio_sample}

\item{allowed_range_npm}{allowed_range_npm c(min_npm, max_npm)}
}
\value{
a dataframe
}
\description{
Prepare the financial data following those steps :
\itemize{
\item Remove implausible values from the raw data, by setting the cell where it appears to NA .
Those values will be filled with averages later on.
\item Using the country associated to each asset, map each country to a world region.
We're not using the scenarios regions, because that would create duplicate rows.
\item When a company owns several assets in different countries, but same region, aggregate the financial indicators from country to region.
\item Add new rows with NA financial value for the missing company_id from companies_data.
\item If the ownership tree is provided as an input, match missing financial values of subsidiaries
to their closest parent compahy, if the parent company has non-NA financial values in the raw data.
\item Compute averages, and pick the most granular one possible to fill missing values
\item As financial indicators are still at the (company_id, ald_sector, region) granularity level,
aggregate those to the company level to be used in the stress test.
}
}
