% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production_types}
\alias{production_types}
\title{Used to generate company synthetic datasets
TODO remove to be}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 4 columns.
}
\source{
Scenario analysis input available production types
}
\usage{
production_types
}
\description{
Used to generate company synthetic datasets
TODO remove to be
}
\examples{
production_types
}
\keyword{datasets}
